Every action now requires smart card authentication

+++ index.php
@@ -2,16 +2,16 @@
 
 <html>
 <head>
+    <script src="sign.html"></script>
+    <script src="sign.js"></script>
     <title>Credit Cards Managment</title>
 </head>
 
-<body>
-    
+<body>   
     <?php
     
     // Add the database connection logic
     include "dbconnect.php";
-    //include "SignUtil/sign.js";
 
     define("ENCRYPTION_KEY", "!@#$%^&*");
 
@@ -24,32 +24,51 @@
         return $encrypted_string;
     }
 
-    
+    function isUserValid()
+    {
+        echo '<script type="text/javascript">'
+            ,'signText(sign);'
+            ,'</script>';
+
+        $result = $_GET['signed'];
+
+        if ($result == 'true')
+            return true;
+        return false;
+    }
+
     // Populates database table fields with INSERT command
     if (isset($_POST['enter'])) 
-    {    
-        // Create a random pin code
-        $pincode = substr(md5(microtime()), rand(0,26), 4);
-        // Encrypt the pin code
-        $encrypted = encrypt($pincode, ENCRYPTION_KEY);
-
-        $sql = "INSERT INTO CreditCards (CardNumber, UserFName, UserLName, PinCode)
-                VALUES ('$_POST[cardnumber]','$_POST[firstname]','$_POST[lastname]','" . $encrypted . "')";
-            
-        if (!mysqli_query($connection,$sql))
+    {   
+        if ($_POST['cardnumber'] != null)
         {
-            die('Error: ' . mysqli_error($connection));
+            if (isUserValid())
+            {
+                // Create a random pin code
+                $pincode = substr(md5(microtime()), rand(0,26), 4);
+                // Encrypt the pin code
+                $encrypted = encrypt($pincode, ENCRYPTION_KEY);
+
+                $sql = "INSERT INTO CreditCards (CardNumber, UserFName, UserLName, PinCode)
+                        VALUES ('$_POST[cardnumber]','$_POST[firstname]','$_POST[lastname]','" . $encrypted . "')";
+                    
+                if (!mysqli_query($connection,$sql))
+                {
+                    die('Error: ' . mysqli_error($connection));
+                }
+            }
         }  
     }
     
     // Delete listing by ID
     if (isset($_POST['deletebutton']))
     {
-        $delete = $_POST['delete'];
-        mysqli_query($connection,"DELETE FROM CreditCards WHERE CardNumber='$delete'");
+        if (isUserValid())
+        {
+            $delete = $_POST['delete'];
+            mysqli_query($connection,"DELETE FROM CreditCards WHERE CardNumber='$delete'");
+        }
     }
-    
-    /*mysqli_close($connection);*/
 
 ?>
     <!--Form fields for inserting a record-->
@@ -57,7 +76,7 @@
         Card Number:    <input type="text" name="cardnumber"><br>
         First Name:     <input type="text" name="firstname"><br>
         Last Name:      <input type="text" name="lastname"><br>
-        <input type="submit" name="enter">
+        <input type="submit" name="enter">        
     </form>
     
     <br>
@@ -84,7 +103,6 @@
     echo "<table border='1'>
     
     <tr>
-    <th>ID</th>
     <th>Card Number</th>
     <th>First Name</th>
     <th>Last Name</th>
@@ -94,7 +112,6 @@
     while ($row = mysqli_fetch_array($table))
     {
         echo "<tr>";
-        echo "<td>" . $row['ID'] . "</td>";
         echo "<td>" . $row['CardNumber'] . "</td>";
         echo "<td>" . $row['UserFName'] . "</td>";
         echo "<td>" . $row['UserLName'] . "</td>";
@@ -106,9 +123,9 @@
 }
 
 ?>
-     <br>
+    <br>
    
-   <!--Button that calls on upadate.php and views it's contents-->
+    <!--Button that calls on upadate.php and views it's contents-->
     <form action="index.php" method="post">
     <input type="submit" value="Upate Record" name="updatebutton">
     </form>
@@ -123,10 +140,16 @@
     
     //updates listing in database by ID // processes update.php file
     
-     if (isset($_POST['updatelisting'])){
-        $update = $_POST['update'];
-    mysqli_query($connection,"UPDATE CreditCards SET CardNumber='$_POST[cardnumber]', UserFName='$_POST[firstname]', UserLName='$_POST[lastname]'
-    WHERE ID='$update'");
+    if (isset($_POST['updatelisting']))
+    {
+        if (isUserValid())
+        {
+            $update = $_POST['update'];
+            mysqli_query($connection,
+                "UPDATE CreditCards SET CardNumber='$_POST[cardnumber]',
+                UserFName='$_POST[firstname]', UserLName='$_POST[lastname]'
+                WHERE ID='$update'");
+        }
     ?>
     
     <!--Script reloads page so updated field can be viewed-->
